==> 3._missingImputation.Rmd <==
---
title: "2._missingImputation"
author: "Divya Prima Crasta-237879"
date: "2025-07-14"
output: html_document
---

```{r}
library(caret)
library(dplyr)

library(mlr3)
library(mlr3tuning)
library(mlr3learners)
library(paradox)

library(miceRanger)
library(missForest)
library(randomForest)
```
```{r}
load('E:\Thesis\Report\Codes\data\combined_assay_data.RData')
X <- assay.data[,3:length(assay.data)]
assay.data$Toxicity <- factor(assay.data$Toxicity, levels = unique(assay.data$Toxicity))
y <- assay.data$Toxicity
```
```{r}
# before logarithmizing
X_n <- apply(X[,2:length(X)], 2, function(x) x / X$Cmax)
X_n_l <- apply(X_n, 2, log)
X_l <- apply(X, 2, log)


X_n_l2 = X_n_l
X_n_l2[penalty.data[,-1] == 1 ] <-  NA
X_n_l2 = data.frame(X_n_l2)
data_n_l2 <- cbind(X_n_l2, Toxicity = as.factor(y))

#X_n_l is normalized by Cmax and logarithmized
#X_n_l2 is normalized by Cmax and logarithmized with missing values
```
```{r}
wholeImputeFun <- function(X, imputeFun, ...){
    set.seed(123)
    k <- 5
    folds <- createFolds(factor(y), k = 5, returnTrain = FALSE)
    # Initialize a vector to store imputed values in the original order
    imputed_X_all <- data.frame(matrix(NA, nrow = nrow(X), ncol = ncol(X)))
    for(i in 1:k){
        test_indices <- folds[[i]]
        train_X <- X[-test_indices, ]
        test_X <- X[test_indices, ]
        imputed_test_X <- do.call(imputeFun, list(train_X, test_X, ...))
        # Store the imputed data in the correct positions
        imputed_X_all[test_indices, ] <- imputed_test_X
    }
    imputed_data_all <- as.data.frame(cbind(Toxicity = y, imputed_X_all))
    task <- TaskClassif$new(id = "data", backend = imputed_data_all, target = "Toxicity")
    return(task)
    }
```
```{r}
miceImpute <- function(train_X, test_X, ...){
    mrModelOutput <- miceRanger(train_X, valueSelector = "value", returnModels = TRUE, verbose = FALSE)
    new_X_test <- impute(test_X, mrModelOutput, verbose = FALSE)
    return(new_X_test$imputedData[[5]])
}
```
```{r}
mice2Impute <- function(train_X, test_X, ...){
    mrMeanMatch <- miceRanger(train_X, valueSelector = "meanMatch", returnModels = TRUE, verbose=FALSE)
    new_X_test <- impute(test_X, mrMeanMatch, verbose = FALSE)
    return(new_X_test$imputedData[[5]])
}
```
```{r}
missForestImpute <- function(train_X, test_X, ...){
    imp_train_X <- missForest(train_X)$ximp
    # 2) combine
    train_test_X <- rbind(test_X, imp_train_X)
    # 3) impute test
    imp_test_X <- missForest(train_test_X)$ximp[1:nrow(test_X), ]
    return(imp_test_X)
    }
```
```{r}
RFImpute <- function(train_X, test_X, y){
    names(y) <- seq(1, 100, 1)
    train_y <- y[rownames(train_X)]
    test_y <- y[rownames(test_X)]
    imp_train_X <- rfImpute(Toxicity~., y = train_y, data = cbind(train_X, Toxicity = train_y)) %>% select(-Toxicity)
    test_train_X <- rbind(test_X, imp_train_X)
    test_train_y <- c(test_y, train_y)
    imp_test_X <- rfImpute(Toxicity~., y = test_train_y, data = cbind(test_train_X, Toxicity = test_train_y))[1:nrow(test_X),] %>% select(-Toxicity)
    return(imp_test_X)
}
```
```{r}
compute_accs <- function(task_m, tuneFun){
    tuneResults <- list()
    for(i in 1:3){
    tuneResults[[i]] <- do.call(tuneFun, list(task_m))
    }
    print(lapply(tuneResults, function(result) result$classif.ce))
    return(tuneResults)
}
```
```{r}
tuneFun_xg <- function(task_m){
    lrn_xgb <- lrn("classif.xgboost",
    nrounds = to_tune(100, 500, logscale = FALSE),
    max_depth = to_tune(1, 5, logscale = FALSE),
    eta = to_tune(0.1, 0.5, logscale = FALSE)
    )
    instance_xgb_m <- ti(
      task = task_m,
      learner = lrn_xgb,
      resampling = rsmp("cv", folds = 5),
      measures = msr("classif.ce"),
      terminator = trm("evals", n_evals = 50)
    )
    tuner = tnr("random_search")
    tuner$optimize(instance_xgb_m)
    return(instance_xgb_m)
}
```
```{r}
task_m1 <- wholeImputeFun(X_n_l2, miceImpute)
# XGBoost nested tuning
results_m1_xg <- nested_tuning(task_m1, tuneFun_xg) 
```
```{r}
task_m2 <- wholeImputeFun(X_n_l2, mice2Impute)
results_m2_xg <- nested_tuning(task_m2, tuneFun_xg)
```
```{r}
task_miss <- wholeImputeFun(X_n_l2, missForestImpute, y)
results_miss_xg <- nested_tuning(task_miss, tuneFun_xg)
```
```{r}
task_rf <- wholeImputeFun(X_n_l2, RFImpute, y)
results_rf_xg <- nested_tuning(task_rf, tuneFun_xg)
```
# Storing imputed values
```{r}
# Imputed Data
# Extract the data from task_m1
X_imp <- task_m1$data()

# Ensure it's a data.table
library(data.table)
setDT(X_imp)  # Converts X_imp to a data.table (if not already)

# Define the desired column order
sorted_colnames <- paste0("X", 1:24)

# Reorder columns correctly
X_imp <- X_imp[, ..sorted_colnames]  # Use .. to reference a variable in data.table

# Check new column order
print(colnames(X_imp))

colnames(X_imp) <- colnames(X_n_l2)

data_imp <- cbind(X_imp, Toxicity = y)
```
```{r}
write.csv(data_imp, "E:\Thesis\Report\Codes\data\data_imputed.csv")
```
